+++ ee/tabby-ui/app/(dashboard)/experiments/components/feature-list.tsx
